♻️ Refactor: 종속형 utils 해당 컴포넌트에 종속
