+++ src/common/components/DefaultProfileImage/DefaultProfileImage.server.tsx
@@ -2,7 +2,7 @@ import clsx from 'clsx';
 
 import { profileColors } from '@common/constants/profile';
 
-import hashCode from '@common/utils/hash-code';
+import hashCode from '@common/components/DefaultProfileImage/utils/hash-code';
 
 import { DefaultProfileImageProps } from './DefaultProfileImage.types';
