+++ src/common/components/DefaultProfileImage/utils/hash-code.ts
